+++ packages/core/database/migrations/2022_03_30_100000_update_quantity_on_order_lines_table.php
@@ -4,18 +4,18 @@ use GetCandy\Base\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class UpdateQuantityOnCartLinesTable extends Migration
+class UpdateQuantityOnOrderLinesTable extends Migration
 {
     public function up()
     {
-        Schema::table($this->prefix.'cart_lines', function (Blueprint $table) {
+        Schema::table($this->prefix.'order_lines', function (Blueprint $table) {
             $table->unsignedInteger('quantity')->change();
         });
     }
 
     public function down()
     {
-        Schema::table($this->prefix.'cart_lines', function ($table) {
+        Schema::table($this->prefix.'order_lines', function ($table) {
             $table->smallInteger('quantity')->unsigned()->change();
         });
     }